requirements.txt
# Real-time Price System Requirements

# Core Framework
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.9

# Database
duckdb==1.3.2

# HTTP Client
httpx
aiohttp==3.12.15

# WebSocket Support
websockets==15.0.1

# Utilities
python-multipart==0.0.20
tenacity==9.1.2

# Async Support
anyio==4.10.0

# Environment
python-dotenv

# Logging (optional, for colored output)
colorama

# Development (optional)
# pytest
# pytest-asyncio
# black
# flake8
